requirements.txt
Flask==2.3.3
python-decouple==3.8
Requests==2.31.0
spotipy==2.23.0
